Change time managing function
