+++ dpmModule/kernel/abstract.py
@@ -100,39 +100,6 @@ class AbstractScenarioGraph():
         return _nodes, _links
 
 
-class AbstractScheduler():
-    def __init__(self, graph):
-        self.graph = graph
-        self.total_time_left= None
-        self.total_time_initial = None
-
-    def initialize(self, time):
-        raise NotImplementedError(''' AbstractScheduler.initializer(time) must be implemented,
-        this function will initialize scheduler with given Schduling total time.
-        ''')
-
-    def get_current_time(self):
-        return self.total_time_initial - self.total_time_left
-
-    def is_simulation_end(self):
-        return (self.total_time_left < 0)
-
-    def spend_time(self, time):
-        '''This function might be overrided, with super().spend_time(time) calling.
-        '''
-        self.total_time_left -= time
-        self.graph.spend_time(time)
-
-    def dequeue(self):
-        raise NotImplementedError('''Scheduler.dequeue() must be implemented,
-        This function will return appropriate element that will be executed by
-        Simulator, which satisfy every rule, for FetchingPolicy.
-        ''')
-
-    def get_delayed_task(self):
-        raise NotImplementedError('''Scheduler.get_delayed_task() must be implemented,
-        This function will return delayed task whlie previous task pending.''')
-
 class AbstractSession():
     pass
